// frontend/src/types/BowlingLeague.ts
export type BowlingLeague = {
  team: any;
  bowlerId: number;
  bowlerLastName: string;
  bowlerFirstName: string;
  bowlerMiddleInit: string;
  bowlerAddress: string;
  bowlerCity: string;
  bowlerState: string;
  bowlerZip: string;
  bowlerPhoneNumber: string;
};

export type Team = {
  teamId: number;
  teamName: string;
};